Compatibility with EMONPI In EMONPI mode, a script is provided which is used to send updates to the display. This uses emonhub to send the data via RF. The advantage is that the script 'emonglcd-send.pl' can be configured to subscribe to any topic using MQTT and use those values sent directly to the display.
This is useful as it avoids having to use EMONTX and 'sniff' the data
off the air. EMONHUB/EMONPI can maintain the solar PV and utility
readings and calculate the Kwh. Any feeds or inputs can be sent to the
display for these values.

This is useful if you are using emonhub for aggregation purposes,
potentially using other sources (modbus reading via external power
meters or reading inverter directly) rather that CT coils connected to
emonpi/emontx.

Any emonhub input or feed can be used to set the values on the display,
and all is configurable in emonglcd-send.cfg